+++ src/migrations/Install.php
@@ -18,6 +18,7 @@ class Install extends Migration
             'sourceSiteId' => $this->integer()->null(),
             'targetId' => $this->integer()->notNull(),
             'sortOrder' => $this->integer()->notNull(),
+            'pinned' => $this->boolean()->notNull()->defaultValue(false),
             'dateCreated' => $this->dateTime()->notNull(),
             'dateUpdated' => $this->dateTime()->notNull(),
             'uid' => $this->uid(),